requirements/base.txt
asyncpg~=0.25
ujson~=5.4
tortoise-orm~=0.19